\name{vudc-package}
\alias{vudc-package}
\alias{vudc}
\docType{package}
\title{
Visualization of Univariate Data for Comparison}
\description{
Contains 2 new functions for visualization of univariate data for comparison: ccdplot and qddplot, along with input data.
}
\details{
\tabular{ll}{
Package: \tab vudc\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-11-15\cr
License: \tab GPL (>= 2)\cr
}
The data visualization functions are the following:
\itemize{
	\item \code{\link{ccdplot}} - Cumulative Characteristid Diagram
	\item \code{\link{qddplot}} - Quantile Differece Diagram
}
The package contains input data about 4 projects written in Java (\code{\link{projectdata}}).
}
\author{
Csaba Farago <farago@inf.u-szeged.hu>
}
\keyword{ package }
