\name{vudc-package}
\alias{vudc-package}
\alias{vudc}
\docType{package}
\title{
Visualization of Univariate Data for Comparison}
\description{
Contains two functions for visualization univariate data: \code{ccdplot} and \code{qddplot}.
}
\details{
\tabular{ll}{
Package: \tab vudc\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2016-02-20\cr
License: \tab GPL (>= 2)\cr
}
The data visualization functions are the following:
\itemize{
	\item \code{\link{ccdplot}} - Cumulative Characteristic Diagram
	\item \code{\link{qddplot}} - Quantile Difference Diagram
}
The package contains input data for case study, related to four software projects written in Java (\code{\link{projectdata}}).
}
\author{
Csaba Farago <farago@inf.u-szeged.hu>
}
\keyword{package}

\references{
  Csaba Farago. Visualization of Univariate Data for Comparison. \emph{Annales Mathematicae et Informaticae}, volume 45, pages 39--53, 2015.
}
